refactor: add agenda folder and move related files in it
