+++ androidApp/src/main/java/com/example/tasky/android/agenda/AgendaGraph.kt
@@ -3,14 +3,14 @@ package com.example.tasky.android.agenda
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.navigation
+import com.example.tasky.agenda.data.model.Event
+import com.example.tasky.agenda.data.model.Reminder
+import com.example.tasky.agenda.data.model.Task
 import com.example.tasky.android.agenda.screen.Agenda
 import com.example.tasky.android.agenda.screen.AgendaDetailsScreenType
 import com.example.tasky.android.agenda.screen.agendaDetailsScreen
 import com.example.tasky.android.agenda.screen.agendaScreen
 import com.example.tasky.android.agenda.screen.navigateToAgendaDetails
-import com.example.tasky.model.agenda.Event
-import com.example.tasky.model.agenda.Reminder
-import com.example.tasky.model.agenda.Task
 import kotlinx.serialization.Serializable
 
 @Serializable
