+++ androidApp/src/main/java/com/example/tasky/android/agenda/components/details/DetailsDeleteSection.kt
@@ -8,13 +8,13 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
+import com.example.tasky.agenda.data.model.AgendaItem
+import com.example.tasky.agenda.data.model.Event
+import com.example.tasky.agenda.data.model.Reminder
+import com.example.tasky.agenda.data.model.Task
 import com.example.tasky.android.R
 import com.example.tasky.android.theme.Gray
 import com.example.tasky.android.theme.MyApplicationTheme
-import com.example.tasky.model.agenda.AgendaItem
-import com.example.tasky.model.agenda.Event
-import com.example.tasky.model.agenda.Reminder
-import com.example.tasky.model.agenda.Task
 
 @Composable
 fun DetailsDeleteSection(
