+++ androidApp/src/main/java/com/example/tasky/android/agenda/di/AgendaDetailsModule.kt
@@ -1,10 +1,10 @@
 package com.example.tasky.android.agenda.di
 
+import com.example.tasky.agenda.domain.AgendaRepository
+import com.example.tasky.agenda.domain.IAgendaRepository
 import com.example.tasky.android.agenda.viewmodel.AgendaDetailsViewModel
 import com.example.tasky.android.utils.IImageCompressor
 import com.example.tasky.android.utils.ImageCompressor
-import com.example.tasky.repository.AgendaRepository
-import com.example.tasky.repository.IAgendaRepository
 import org.koin.core.module.dsl.factoryOf
 import org.koin.core.module.dsl.singleOf
 import org.koin.core.module.dsl.viewModelOf
