+++ androidApp/src/main/java/com/example/tasky/android/agenda/di/AgendaModule.kt
@@ -1,8 +1,8 @@
 package com.example.tasky.android.agenda.di
 
+import com.example.tasky.agenda.domain.AgendaRepository
+import com.example.tasky.agenda.domain.IAgendaRepository
 import com.example.tasky.android.agenda.viewmodel.AgendaViewModel
-import com.example.tasky.repository.AgendaRepository
-import com.example.tasky.repository.IAgendaRepository
 import org.koin.core.module.dsl.singleOf
 import org.koin.core.module.dsl.viewModelOf
 import org.koin.dsl.bind
