+++ androidApp/src/main/java/com/example/tasky/android/agenda/screen/AgendaDetailsScreen.kt
@@ -29,6 +29,11 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
+import com.example.tasky.agenda.data.model.AgendaItem
+import com.example.tasky.agenda.data.model.Event
+import com.example.tasky.agenda.data.model.Photo
+import com.example.tasky.agenda.data.model.Reminder
+import com.example.tasky.agenda.data.model.Task
 import com.example.tasky.android.R
 import com.example.tasky.android.agenda.components.details.DetailsAttendeeSection
 import com.example.tasky.android.agenda.components.details.DetailsAttendeeSectionTabOption
@@ -51,11 +56,6 @@ import com.example.tasky.android.theme.Light
 import com.example.tasky.android.theme.MyApplicationTheme
 import com.example.tasky.android.utils.ObserveAsEvents
 import com.example.tasky.android.utils.serializableNavType
-import com.example.tasky.model.agenda.AgendaItem
-import com.example.tasky.model.agenda.Event
-import com.example.tasky.model.agenda.Photo
-import com.example.tasky.model.agenda.Reminder
-import com.example.tasky.model.agenda.Task
 import kotlinx.collections.immutable.toImmutableList
 import kotlinx.datetime.Instant
 import kotlinx.datetime.LocalDate
