+++ androidApp/src/main/java/com/example/tasky/android/agenda/viewmodel/AgendaDetailsViewModel.kt
@@ -5,6 +5,14 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import androidx.navigation.toRoute
+import com.example.tasky.agenda.data.model.Attendee
+import com.example.tasky.agenda.data.model.CreateEventBody
+import com.example.tasky.agenda.data.model.Event
+import com.example.tasky.agenda.data.model.Reminder
+import com.example.tasky.agenda.data.model.Task
+import com.example.tasky.agenda.data.model.UpdateEventBody
+import com.example.tasky.agenda.data.model.copy
+import com.example.tasky.agenda.domain.IAgendaRepository
 import com.example.tasky.android.agenda.components.details.DetailsEditTextType
 import com.example.tasky.android.agenda.components.details.DetailsPhoto
 import com.example.tasky.android.agenda.screen.AgendaDetails
@@ -13,20 +21,11 @@ import com.example.tasky.android.agenda.screen.AgendaDetailsScreenState
 import com.example.tasky.android.agenda.screen.AgendaDetailsScreenType
 import com.example.tasky.android.utils.IImageCompressor
 import com.example.tasky.manager.SessionManager
-import com.example.tasky.model.agenda.Attendee
-import com.example.tasky.model.agenda.CreateEventBody
-import com.example.tasky.model.agenda.Event
-import com.example.tasky.model.agenda.Reminder
-import com.example.tasky.model.agenda.Task
-import com.example.tasky.model.agenda.UpdateEventBody
-import com.example.tasky.model.agenda.copy
 import com.example.tasky.model.onSuccess
-import com.example.tasky.repository.IAgendaRepository
 import com.example.tasky.util.toLocalDateTime
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
