+++ androidApp/src/main/java/com/example/tasky/android/agenda/viewmodel/AgendaViewModel.kt
@@ -2,15 +2,15 @@ package com.example.tasky.android.agenda.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.example.tasky.agenda.data.model.AgendaItem
+import com.example.tasky.agenda.data.model.Task
+import com.example.tasky.agenda.domain.IAgendaRepository
 import com.example.tasky.android.agenda.screen.AgendaItemUi
 import com.example.tasky.android.agenda.screen.AgendaScreenEvent
 import com.example.tasky.android.agenda.screen.AgendaScreenState
 import com.example.tasky.manager.SessionManager
 import com.example.tasky.manager.loginManager
-import com.example.tasky.model.agenda.AgendaItem
-import com.example.tasky.model.agenda.Task
 import com.example.tasky.model.onSuccess
-import com.example.tasky.repository.IAgendaRepository
 import com.example.tasky.util.getAvatarDisplayName
 import kotlinx.collections.immutable.toImmutableList
 import kotlinx.coroutines.flow.MutableStateFlow
