+++ shared/src/commonMain/kotlin/com/example/tasky/agenda/data/AgendaDataSource.kt
@@ -1,22 +1,22 @@
-package com.example.tasky.dataSource
+package com.example.tasky.agenda.data
 
+import com.example.tasky.agenda.data.model.Agenda
+import com.example.tasky.agenda.data.model.CreateEventBody
+import com.example.tasky.agenda.data.model.CreateReminderBody
+import com.example.tasky.agenda.data.model.CreateTaskBody
+import com.example.tasky.agenda.data.model.Event
+import com.example.tasky.agenda.data.model.EventPath
+import com.example.tasky.agenda.data.model.GetAgendaResponse
+import com.example.tasky.agenda.data.model.Reminder
+import com.example.tasky.agenda.data.model.ReminderPath
+import com.example.tasky.agenda.data.model.Task
+import com.example.tasky.agenda.data.model.TaskPath
+import com.example.tasky.agenda.data.model.UpdateEventBody
+import com.example.tasky.agenda.data.model.UpdateReminderBody
+import com.example.tasky.agenda.data.model.UpdateTaskBody
 import com.example.tasky.manager.HttpManager
 import com.example.tasky.model.BaseError
 import com.example.tasky.model.ResultWrapper
-import com.example.tasky.model.agenda.Agenda
-import com.example.tasky.model.agenda.CreateEventBody
-import com.example.tasky.model.agenda.CreateReminderBody
-import com.example.tasky.model.agenda.CreateTaskBody
-import com.example.tasky.model.agenda.Event
-import com.example.tasky.model.agenda.EventPath
-import com.example.tasky.model.agenda.GetAgendaResponse
-import com.example.tasky.model.agenda.Reminder
-import com.example.tasky.model.agenda.ReminderPath
-import com.example.tasky.model.agenda.Task
-import com.example.tasky.model.agenda.TaskPath
-import com.example.tasky.model.agenda.UpdateEventBody
-import com.example.tasky.model.agenda.UpdateReminderBody
-import com.example.tasky.model.agenda.UpdateTaskBody
 import com.example.tasky.util.safeCall
 import io.ktor.client.HttpClient
 import io.ktor.client.plugins.resources.delete
